퀴즈 관련 이용자 비교 로직 작성

+++ finalproject/src/main/java/com/kh/finalproject/restcontroller/QuizRestController.java
@@ -9,12 +9,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.kh.finalproject.dto.QuizDto;
+import com.kh.finalproject.error.NeedPermissionException;
+import com.kh.finalproject.error.TargetNotfoundException;
 import com.kh.finalproject.service.QuizService;
+import com.kh.finalproject.vo.TokenVO;
 
 @CrossOrigin
 @RestController
@@ -28,11 +32,12 @@ public class QuizRestController {
 	//TokenVO로 사용자 조회 및 quiz_creator_id로 등록
 	@PostMapping("/")
 	public QuizDto insert(
-			//TokenVO 추가 예정
-			@RequestBody QuizDto quizDto
+			@RequestBody QuizDto quizDto,
+			@RequestAttribute TokenVO tokenVO
 			) {
-		//토큰에서 memberId 추출 한 후
-		//quizDto.setQuizCreatorId(memberId);
+		//토큰에서 memberId 추출 한 후 작성자 설정
+		String memberId = tokenVO.getLoginId();
+		quizDto.setQuizCreatorId(memberId);
 		
 		return quizService.registQuiz(quizDto);
 	}
@@ -41,9 +46,13 @@ public class QuizRestController {
 	//TokenVO로 사용자 조회
 	@GetMapping("/game/{contentsId}")
 	public List<QuizDto> game(
-			//TokenVO 추가 예정
+			@RequestAttribute TokenVO tokenVO,
 			@PathVariable long contentsId
 			){
+		
+		//비회원은 나가라만 설정해도 되나?
+		if(tokenVO == null) throw new TargetNotfoundException();
+		
 		
 		return quizService.getQuizGame(contentsId, "TokenVO.getLoginId()");
 	}
@@ -51,14 +60,26 @@ public class QuizRestController {
 	//해당 영화의 퀴즈 목록 조회 구문
 	//관리자 페이지로 뺄 가능성 높음
 	@GetMapping("/list/{contentsId}")
-	public List<QuizDto> getList(@PathVariable long contentsId) {
-		//관리자 권한 체크 로직 필요
+	public List<QuizDto> getList(
+			@RequestAttribute TokenVO tokenVO,
+			@PathVariable long contentsId
+			) {
+		
+		//관리자 권한 체크 로직
+		boolean isAdmin = tokenVO.getLoginLevel().equals("관리자");
+		if(isAdmin == false) throw new NeedPermissionException();
+		
 		return quizService.getQuizList(contentsId);
 	}
 	
 	//퀴즈 상세정보 조회
 	@GetMapping("/{quizId}")
 	public QuizDto detail(@PathVariable long quizId) {
+		
+		//quizId로 조회한 퀴즈가 없다면 나가는 코드 작성?
+		QuizDto findDto = quizService.getQuizDetail(quizId);
+		if(findDto == null) throw new TargetNotfoundException();
+		
 		return quizService.getQuizDetail(quizId);
 	}
 	
@@ -66,19 +87,35 @@ public class QuizRestController {
 	//TokenVO로 사용자 조회
 	@PatchMapping("/")
 	public boolean update(
-			//TokenVO 추가 예정,
+			@RequestAttribute TokenVO tokenVO,
 			@RequestBody QuizDto quizDto) {
 		
-		//토큰 ID와 quizDto.getCreatorId()가 일치하는지 확인
+		//로그인 한 계정과 작성자의 아이디가 같은지 검사
+		String loginId = tokenVO.getLoginId();
+		String creatorId = quizDto.getQuizCreatorId();
+		
+		boolean isCorrect = loginId.equals(creatorId);
+		if(isCorrect) throw new NeedPermissionException();
+		
 		return quizService.editQuiz(quizDto);
 	}
 	
-	//퀴즈 상태 변경(BLIND, DELETED)
-	//삭제 구문 대신 사용할 메소드
-	//본인 확인 or 관리자인지 확인하는 구문 추가 예정
+	//퀴즈 상태 변경(BLIND)
 	@PatchMapping("/status")
-	public boolean changeStatus(@RequestBody QuizDto quizDto) {
+	public boolean changeStatus(
+			@RequestAttribute TokenVO tokenVO,
+			@RequestBody QuizDto quizDto) {
+		
 		// quizDto에는 quizId와 quizStatus("BLIND")가 들어있어야 함
+		String loginId = tokenVO.getLoginId();
+		String loginLevel = tokenVO.getLoginLevel();
+		
+		//작성자 or 관리자인지 검사
+		boolean isOwner = loginId.equals(quizDto.getQuizCreatorId());
+		boolean isAdmin = loginLevel.equals("관리자");
+		
+		if(isOwner || isAdmin) throw new NeedPermissionException();
+		
 		return quizService.changeQuizStatus(quizDto);
 	}
 	
@@ -92,11 +129,21 @@ public class QuizRestController {
 	//퀴즈 삭제
 	@DeleteMapping("/{quizId}")
 	public boolean delete(
-			//TokenVO 추가 예정,
+			@RequestAttribute TokenVO tokenVO,
 			@PathVariable long quizId) {
 		
 		//TokenVO에서 loginId와 loginLevel 추출해서
-		//본인 확인 or 관리자 권한 확인 구현 예정
+		String loginId = tokenVO.getLoginId();
+		String loginLevel = tokenVO.getLoginLevel();
+		
+		//퀴즈가 있는지 검사
+		QuizDto findDto = quizService.getQuizDetail(quizId);
+		
+		//작성자 or 관리자인지 검사
+		boolean isOwner = loginId.equals(findDto.getQuizCreatorId());
+		boolean isAdmin = loginLevel.equals("관리자");
+		
+		if(isOwner || isAdmin) throw new NeedPermissionException();
 		
 		return quizService.deleteQuiz(quizId);
 	}
